Moved NodeUpdates out of index provider API
- removed unused LabelSchemaSupplier.Box
- moved PropertyLoader out of index provider API
